services/product-service/requirements.txt
amqp==5.3.1
asgiref==3.9.1
billiard==4.2.1
celery==5.3.4
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
dj-database-url==2.1.0
Django==5.2.5
django-cors-headers==4.3.1
django-filter==25.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
idna==3.10
kombu==5.3.4
prompt_toolkit==3.0.52
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
redis==5.0.1
requests==2.31.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
